refactor: move `ui` -> `features/account/delete`
