+++ frontend/src/components/features/account/delete/DeleteAccount.tsx
@@ -1,9 +1,8 @@
 'use client'
 
 import { useMediaQuery } from '~/hooks/useMediaQuery'
-
-import { DeleteAccountDialog } from './dialogs/DeleteAccountDialog'
-import { DeleteAccountDrawer } from './drawers/DeleteAccountDrawer'
+import { DeleteAccountDialog } from './DeleteAccountDialog'
+import { DeleteAccountDrawer } from './DeleteAccountDrawer'
 
 export function DeleteAccount() {
   const isDesktop = useMediaQuery('(min-width: 768px)')
