+++ frontend/src/components/features/account/delete/DeleteAccountDrawer.tsx
@@ -4,36 +4,37 @@ import { useAtomValue } from 'jotai'
 import { useState } from 'react'
 import { userStateAtom } from '~/state/user-state.atom'
 import { getDeleteDescription } from '~/utils/get-delete-description'
-import { Button } from '../buttons/Button'
-import DeleteAccountForm from '../forms/DeleteAccountForm'
-import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogTrigger } from './Dialog'
+import { Button } from '../../../ui/buttons/Button'
+import { Drawer, DrawerContent, DrawerHeader, DrawerTitle, DrawerTrigger } from '../../../ui/drawers/Drawer'
+import DeleteAccountForm from './DeleteAccountForm'
 
-export function DeleteAccountDialog() {
+export function DeleteAccountDrawer() {
   const [open, setOpen] = useState(false)
   const user = useAtomValue(userStateAtom)
+
   return (
-    <Dialog open={open} onOpenChange={setOpen}>
-      <DialogTrigger asChild>
+    <Drawer open={open} onOpenChange={setOpen}>
+      <DrawerTrigger asChild>
         <Button
           className="border border-destructive bg-white text-destructive hover:bg-destructive hover:text-white"
         >
           削除する
         </Button>
-      </DialogTrigger>
-      <DialogContent className="sm:max-w-[470px]">
-        <DialogHeader>
-          <DialogTitle>
+      </DrawerTrigger>
+      <DrawerContent className="mb-8">
+        <DrawerHeader>
+          <DrawerTitle className="text-left">
             アカウントを削除しようとしています
-          </DialogTitle>
-        </DialogHeader>
-        <p className="text-sm leading-6">
+          </DrawerTitle>
+        </DrawerHeader>
+        <p className="pb-4 text-sm leading-6">
           {getDeleteDescription(user.provider)}
           <br />
           この操作は取り消すことができません。
           {user.provider === 'email' && '確認のため、登録中のメールアドレスを入力してください。'}
         </p>
         <DeleteAccountForm onClose={() => setOpen(false)} />
-      </DialogContent>
-    </Dialog>
+      </DrawerContent>
+    </Drawer>
   )
 }
